requirements.txt
Django==1.8.4
Fabric==1.10.2
Markdown==2.6.2
django-formtools==1.0
django-redis-cache==1.6.3
django-redis-sessions==0.5.0
djangorestframework==3.2.4
django-widget-tweaks==1.4.1
gunicorn==19.3.0
hiredis==0.2.0
nose==1.3.7
python-ldap==2.4.21
pexpect==4.0.1
reportlab==3.2.0
psycopg2==2.6.1
redis==2.10.3
six==1.10.0
icalendar==3.9.0
Sphinx==1.3.1
django-debug-toolbar==1.4
django-debug-toolbar-line-profiler==0.4.0
django-extensions==1.5.7
sphinx-bootstrap-theme==0.4.7
Werkzeug==0.10.4
requests==2.8.0
requests-oauthlib==0.5.0
oauthlib==1.0.3
pysftp==0.2.8
pycrypto==2.6.1
django-maintenancemode==0.11.0
bleach==1.4.2
django-cors-headers==1.1.0
django-cacheops==2.4.3
